split more and address CR
