+++ libc/startup/linux/x86_64/CMakeLists.txt
@@ -1,13 +1,27 @@
 add_startup_object(
-  start
+  tls
   SRC
-    start.cpp
+    tls.cpp
   DEPENDS
     libc.config.linux.app_h
     libc.include.sys_mman
     libc.include.sys_syscall
     libc.src.__support.OSUtil.osutil
     libc.src.string.memory_utils.inline_memcpy
+  COMPILE_OPTIONS
+    -fno-stack-protector
+    -fno-omit-frame-pointer
+    -ffreestanding
+    -fno-builtin
+)
+
+add_startup_object(
+  start
+  SRC
+    start.cpp
+  DEPENDS
+    libc.config.linux.app_h
+    libc.src.__support.macros.attributes
   COMPILE_OPTIONS
     -fno-stack-protector
     -fno-omit-frame-pointer
